new update

+++ Controllers/ConfigurationController.cs
@@ -1,4 +1,5 @@
-﻿using DATN.DataContext;
+﻿using System;
+using DATN.DataContext;
 using DATN.Model;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -12,12 +13,13 @@ namespace DATN.Controllers
         private readonly MyDbContext db = context;
 
         [HttpGet()]
-        [Route("getprojecttypelist")]
-        public ProjectTypeModelResponse GetProjectTypeList()
+        [Route("getprojecttypelist/{currentPage?}")]
+        public ProjectTypeModelResponse GetProjectTypeList(int? currentPage)
         {
             var dbS = db.Students;
+            var dbT = db.ProjectType;
             var list =
-                from r in db.ProjectType
+                from r in dbT.Skip(((currentPage ?? 0) - 1) * 10).Take((currentPage ?? 0) * 10)
                 join s in dbS on r.CreatedBy equals s.StudentId into newdbR
                 from newR in newdbR.DefaultIfEmpty()
                 select new ProjectTypeTableModel
@@ -27,7 +29,30 @@ namespace DATN.Controllers
                     Description = r.Description,
                     CreatedBy = newR.StudentName,
                     CreatedDate = r.CreatedDate,
+                    IsWeeklyReport = r.IsWeeklyReport,
                 };
+            return new ProjectTypeModelResponse
+            {
+                ProjectTypeList = [.. list],
+                IsDone = true,
+                Total = dbT.Count(),
+            };
+        }
+
+        [HttpGet()]
+        [Route("getprojecttypeidlist/{status?}")]
+        public ProjectTypeModelResponse GetProjectTypeIdList(string? status)
+        {
+            IQueryable<ProjectTypeTableModel> list;
+            if (status == "0")
+            {
+                list = db.ProjectType;
+            }
+            else
+            {
+                var isWeekly = status == "true";
+                list = db.ProjectType.Where(t => t.IsWeeklyReport == isWeekly);
+            }
             return new ProjectTypeModelResponse { ProjectTypeList = [.. list], IsDone = true };
         }
 
@@ -35,21 +60,44 @@ namespace DATN.Controllers
         [Route("createprojecttype")]
         public ResponseModel CreateProjectType([FromBody] ProjectTypeTableModel model)
         {
-            var sameName = db.ProjectType.Any(item =>
+            ProjectTypeTableModel? pt = null;
+            if (model.ProjectTypeId != null)
+            {
+                pt = db.ProjectType.FirstOrDefault(t => t.ProjectTypeId == model.ProjectTypeId);
+            }
+            var sameName = db.ProjectType.FirstOrDefault(item =>
                 item.ProjectTypeName == model.ProjectTypeName
             );
-            if (sameName)
+            if (pt == null)
             {
-                return new ResponseModel
+                if (sameName != null)
                 {
-                    IsDone = false,
-                    Error = "There is a project type with the same name!",
-                    ErrorCode = 101,
-                };
+                    return new ResponseModel
+                    {
+                        IsDone = false,
+                        Error = "There is a project type with the same name!",
+                        ErrorCode = 101,
+                    };
+                }
+                model.ProjectTypeId = Guid.NewGuid().ToString();
+                db.ProjectType.Add(model);
             }
-
-            model.ProjectTypeId = Guid.NewGuid().ToString();
-            db.ProjectType.Add(model);
+            else
+            {
+                if (sameName?.ProjectTypeId != model.ProjectTypeId)
+                {
+                    return new ResponseModel
+                    {
+                        IsDone = false,
+                        Error = "There is a project type with the same name!",
+                        ErrorCode = 101,
+                    };
+                }
+                pt.ProjectTypeName = model.ProjectTypeName;
+                pt.Description = model.Description;
+                pt.CreatedDate = model.CreatedDate;
+                pt.CreatedBy = model.CreatedBy;
+            }
             try
             {
                 db.SaveChanges();
@@ -69,31 +117,40 @@ namespace DATN.Controllers
             var typeDel = db.ProjectType.FirstOrDefault(item => item.ProjectTypeId == typeId);
             if (typeDel == null)
             {
-                return new ResponseModel { Id = typeId, Error = "Project Type does not exsist !" };
+                return new ResponseModel { Id = typeId, Error = "Project Type does not exsist!" };
             }
-            else
+            var pUseType = db.Projects.FirstOrDefault(item =>
+                item.ProjectTypeId == typeDel.ProjectTypeId
+            );
+            if (pUseType != null)
             {
-                db.ProjectType.Remove(typeDel);
-                try
+                return new ResponseModel
                 {
-                    db.SaveChanges();
-                }
-                catch (DbUpdateException ex)
-                {
-                    return new ResponseModel { IsDone = false, Error = ex.Message };
-                }
-
-                return new ResponseModel { Id = typeId, IsDone = true };
+                    Id = typeId,
+                    Error = "There are a project is using this type!",
+                };
             }
+            db.ProjectType.Remove(typeDel);
+            try
+            {
+                db.SaveChanges();
+            }
+            catch (DbUpdateException ex)
+            {
+                return new ResponseModel { IsDone = false, Error = ex.Message };
+            }
+
+            return new ResponseModel { Id = typeId, IsDone = true };
         }
 
         [HttpGet()]
-        [Route("getstudentrolelist")]
-        public StudentRoleModelResponse GetStudentRoleList()
+        [Route("getstudentrolelist/{currentPage?}")]
+        public StudentRoleModelResponse GetStudentRoleList(int? currentPage)
         {
             var dbS = db.Students;
+            var dbR = db.StudentRole;
             var list =
-                from r in db.StudentRole
+                from r in dbR.Skip(((currentPage ?? 0) - 1) * 10).Take((currentPage ?? 0) * 10)
                 join s in dbS on r.CreatedBy equals s.StudentId into newdbR
                 from newR in newdbR.DefaultIfEmpty()
                 select new StudentRoleTableModel
@@ -104,28 +161,68 @@ namespace DATN.Controllers
                     CreatedBy = newR.StudentName,
                     CreatedDate = r.CreatedDate,
                 };
-            return new StudentRoleModelResponse { StudentRoleList = [.. list], IsDone = true };
+            return new StudentRoleModelResponse
+            {
+                StudentRoleList = [.. list],
+                IsDone = true,
+                Total = dbR.Count(),
+            };
+        }
+
+        [HttpGet()]
+        [Route("getstudentroleidlist")]
+        public StudentRoleModelResponse GetStudentRoleIdList()
+        {
+            return new StudentRoleModelResponse
+            {
+                StudentRoleList = [.. db.StudentRole],
+                IsDone = true,
+            };
         }
 
         [HttpPost()]
         [Route("createstudentrole")]
         public ResponseModel CreateStudentRole([FromBody] StudentRoleTableModel model)
         {
-            var sameName = db.StudentRole.Any(item =>
+            StudentRoleTableModel? sr = null;
+            if (model.StudentRoleId != null)
+            {
+                sr = db.StudentRole.FirstOrDefault(t => t.StudentRoleId == model.StudentRoleId);
+            }
+            var sameName = db.StudentRole.FirstOrDefault(item =>
                 item.StudentRoleName == model.StudentRoleName
             );
-            if (sameName)
+            if (sr == null)
             {
-                return new ResponseModel
+                if (sameName != null)
                 {
-                    IsDone = false,
-                    Error = "There is a student role with the same name!",
-                    ErrorCode = 101,
-                };
-            }
+                    return new ResponseModel
+                    {
+                        IsDone = false,
+                        Error = "There is a student role with the same name!",
+                        ErrorCode = 101,
+                    };
+                }
 
-            model.StudentRoleId = Guid.NewGuid().ToString();
-            db.StudentRole.Add(model);
+                model.StudentRoleId = Guid.NewGuid().ToString();
+                db.StudentRole.Add(model);
+            }
+            else
+            {
+                if (sameName?.StudentRoleId != model.StudentRoleId)
+                {
+                    return new ResponseModel
+                    {
+                        IsDone = false,
+                        Error = "There is a project type with the same name!",
+                        ErrorCode = 101,
+                    };
+                }
+                sr.StudentRoleName = model.StudentRoleName;
+                sr.Description = model.Description;
+                sr.CreatedDate = model.CreatedDate;
+                sr.CreatedBy = model.CreatedBy;
+            }
             try
             {
                 db.SaveChanges();
@@ -149,6 +246,17 @@ namespace DATN.Controllers
             }
             else
             {
+                var sUseType = db.Students.FirstOrDefault(item =>
+                    item.StudentRole == roleDel.StudentRoleId
+                );
+                if (sUseType != null)
+                {
+                    return new ResponseModel
+                    {
+                        Id = roleId,
+                        Error = "There are a student is using this role!",
+                    };
+                }
                 db.StudentRole.Remove(roleDel);
                 try
                 {
